requirements.txt
fastapi==0.115.0
langchain==0.3.0
langchain_core==0.3.5
langchain_openai==0.2.0
langchain_pinecone==0.2.0
openai==1.47.1
pandas==2.2.3
pinecone==5.3.1
pinecone_text==0.9.0
pydantic==2.9.2
python-dotenv==1.0.1
streamlit==1.38.0
uvicorn==0.30.6
langchain-community>=0.0.1
